Proposal of addition to Illegal Immigrant Arrest Rates
Suggestion to make the project that compares illegal immigrant arrest rates a tad more political. 
A big part of Trumps campaign was to build a wall. 
Seems a bit insensitive if you are not someone suffering from the globalization of meritocracy.
It is a fact that immigrants take low paying jobs from these individuals.
pull data to find out how much weight this had in the outcome of the election(possibly a start to an election prediction program).
Very simple project that compares the number of trump voters in red counties to the number of blue collar jobs in those areas.
On the flip side, it will compare the number of Hillary supporters in blue counties to the number of merit-based jobs.
Next iterations could include other touchy, but interesting political stands